Update Map2SVG.py
In Map2SVG.py, the color mapping skips the svg_line if the 'id' attribute is on the same line as the 'style' attribute in the svg file.
I don't see the need for an 'elif' instead of an 'if' statement (all the examples work fine with 'if' instead of 'elif'). 
Otherwise, if it actually is important, the user guide should state explicitly in the 'Appendix: Map Characteristics' section that the 'style' attribute should be on a new line below the 'id' attribute.